+++ packages/instantsearch.js/src/connectors/hierarchical-menu/__tests__/connectHierarchicalMenu-test.ts
@@ -1391,6 +1391,46 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/hierarchica
       ).toWarnDev();
     });
 
+    it('warns when the root of `hierarchicalFacets` is used with conjunctive facets and does not change `SearchParameters`', () => {
+      const render = () => {};
+      const makeWidget = connectHierarchicalMenu(render);
+      const helper = algoliasearchHelper(createSearchClient(), '', {
+        facets: ['category'],
+      });
+
+      const widget = makeWidget({
+        attributes: ['category', 'sub_category'],
+        rootPath: 'TopLevel',
+      });
+
+      expect(() => {
+        const searchParams = widget.getWidgetSearchParameters(helper.state, {
+          uiState: {},
+        });
+        expect(searchParams.hierarchicalFacets).toHaveLength(0);
+      }).toWarnDev();
+    });
+
+    it('warns when the root of `hierarchicalFacets` is used with disjunctive facets and does not change `SearchParameters`', () => {
+      const render = () => {};
+      const makeWidget = connectHierarchicalMenu(render);
+      const helper = algoliasearchHelper(createSearchClient(), '', {
+        disjunctiveFacets: ['category'],
+      });
+
+      const widget = makeWidget({
+        attributes: ['category', 'sub_category'],
+        rootPath: 'TopLevel',
+      });
+
+      expect(() => {
+        const searchParams = widget.getWidgetSearchParameters(helper.state, {
+          uiState: {},
+        });
+        expect(searchParams.hierarchicalFacets).toHaveLength(0);
+      }).toWarnDev();
+    });
+
     describe('with `maxValuesPerFacet`', () => {
       test('returns the `SearchParameters` with default `limit`', () => {
         const render = () => {};