+++ packages/react-instantsearch-hooks-web/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.42.2](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-web@6.42.1...react-instantsearch-hooks-web@6.42.2) (2023-04-11)
+
+**Note:** Version bump only for package react-instantsearch-hooks-web
+
+
+
+
+
+## [6.42.1](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-web@6.42.0...react-instantsearch-hooks-web@6.42.1) (2023-03-28)
+
+**Note:** Version bump only for package react-instantsearch-hooks-web
+
+
+
+
+
+## [6.42.0](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-web@6.41.0...react-instantsearch-hooks-web@6.42.0) (2023-03-21)
+
+
+### Bug Fixes
+
+* **searchbox:** add aria-hidden to svg icons ([#5547](https://github.com/algolia/instantsearch.js/issues/5547)) ([50344e3](https://github.com/algolia/instantsearch.js/commit/50344e3b14c22c886415c0e7d799aca778dc39ab)), closes [#5546](https://github.com/algolia/instantsearch.js/issues/5546)
+
+
+
+
+
 ## [6.41.0](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-hooks-web@6.40.1...react-instantsearch-hooks-web@6.41.0) (2023-03-07)
 
 **Note:** Version bump only for package react-instantsearch-hooks-web